ErrorManager: direct copy logArguments
